+++ redis/redis-core/src/test/java/com/ctrip/xpipe/redis/core/route/impl/Crc32HashRouteChooseStrategyTest.java
@@ -17,7 +17,7 @@ public class Crc32HashRouteChooseStrategyTest {
     @Before
     public void beforeCrc32HashRouteChooseStrategyTest() {
         routes = new LinkedList<>();
-        for(int i = 0; i < 100; i++) {
+        for(long i = 0; i < 100; i++) {
             routes.add(new RouteMeta().setId(i));
         }
     }
